Separate Api controller from views
