+++ app/Http/Controllers/Api/MessagesController.php
@@ -1,13 +1,12 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Api;
 
 use App\Enums\MessageStatus;
 use App\Http\Repositories\MessageRepository;
 use App\Http\Requests\Message\MessageStoreRequest;
 use App\Models\Message;
 use App\Models\Provider;
-use Illuminate\Contracts\View\View;
 use Illuminate\Http\JsonResponse;
 
 class MessagesController extends ResponseController
@@ -17,12 +16,6 @@ class MessagesController extends ResponseController
     public function __construct()
     {
         $this->messageRepository = new MessageRepository(new Message());
-    }
-
-    public function index(): view
-    {
-        $messages = $this->messageRepository->paginate_cache();
-        return view('pages.messages', compact('messages'));
     }
 
     public function store(MessageStoreRequest $request): JsonResponse
