+++ routes/web.php
@@ -17,4 +17,4 @@ Route::get('/', function () {
     return view('app');
 });
 
-Route::get('/messages', [\App\Http\Controllers\MessagesController::class,'index'])->name('messages.index');
+Route::get('/messages', [\App\Http\Controllers\View\MessagesController::class,'index'])->name('messages.index');
